+++ app/concepts/stellar_base/withdrawal_requests/contracts/create.rb
@@ -7,16 +7,17 @@ module StellarBase
         property :asset_code
         property :dest
         property :dest_extra
-        property :fee_network
+        property :fee_fixed
 
         validates(
           :asset_type,
           :asset_code,
           :dest,
-          :fee_network,
+          :fee_fixed,
           presence: true,
         )
         validate :check_valid_asset_code
+        validates(:fee_fixed, numericality: {greater_than_or_equal_to: 0})
 
         def check_valid_asset_code
           asset_codes = StellarBase.configuration.withdrawable_assets