+++ spec/concepts/stellar_base/withdrawal_requests/contracts/create_spec.rb
@@ -9,13 +9,24 @@ module StellarBase
             asset_type
             asset_code
             dest
-            fee_network
+            fee_fixed
           ].each do |attr|
             it "requires #{attr}" do
               contract = described_class.new(WithdrawalRequest.new)
               contract.validate(attr => "")
               expect(contract.errors[attr]).to include("can't be blank")
             end
+          end
+
+          it "fee_fixed must be >= 0" do
+            contract = described_class.new(WithdrawalRequest.new)
+            contract.validate(fee_fixed: -1)
+            expect(contract.errors[:fee_fixed]).
+              to include("must be greater than or equal to 0")
+
+            contract = described_class.new(WithdrawalRequest.new)
+            contract.validate(fee_fixed: 0)
+            expect(contract.errors[:fee_fixed]).to be_empty
           end
 
           context "valid asset_code" do