+++ spec/dummy/app/services/get_withdraw_fee_fixed_from.rb
@@ -1,0 +1,7 @@
+class GetWithdrawFeeFixedFrom
+
+  def self.call(params, asset_details)
+    0.911
+  end
+
+end