+++ spec/dummy/app/services/get_withdraw_fee_fixed_quote_from.rb
@@ -1,4 +1,4 @@
-class GetWithdrawFixedFeeQuoteFrom
+class GetWithdrawFeeFixedQuoteFrom
 
   def self.call(fee_request)
     0.0001