+++ spec/services/stellar_base/find_asset_details_spec.rb
@@ -44,9 +44,9 @@ module StellarBase
           network: "bitcoin",
           asset_code: "BTCT",
           issuer: ENV["ISSUER_ADDRESS"],
-          fee_fixed: 0.01,
           max_amount_from: GetMaxAmount.to_s,
-          fixed_fee_quote_from: GetWithdrawFixedFeeQuoteFrom,
+          fee_fixed_quote_from: GetWithdrawFeeFixedQuoteFrom,
+          fee_fixed_from: GetWithdrawFeeFixedFrom,
         })
       end
     end